requirement.txt
python-jenkins==1.4.0

